:recycle: Split when steps into modules
