+++ test/steps/mnemonic/2_when.js
@@ -1,0 +1,28 @@
+/*
+ * LiskHQ/lisky
+ * Copyright © 2017 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ *
+ */
+import {
+	createMnemonicPassphrase,
+	isValidMnemonicPassphrase,
+} from '../../../src/utils/mnemonic';
+
+export function theMnemonicPassphraseIsValidated() {
+	const { mnemonicPassphrase } = this.test.ctx;
+	this.test.ctx.returnValue = isValidMnemonicPassphrase(mnemonicPassphrase);
+}
+
+export function aNewMnemonicPassphraseIsCreated() {
+	this.test.ctx.mnemonicPassphrase = createMnemonicPassphrase();
+}
